# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
...
  handlers.SearchRequestBody:
    properties:
      collection:
        type: string
      filter:
        additionalProperties: true
        type: object
      min_score:
        type: number
      query:
        type: string
      top_k:
        type: integer
      use_cache:
        type: boolean
    type: object
...
  services.SearchResponse:
    properties:
      collection:
        type: string
      embedding_time_ms:
        type: number
      from_cache:
        type: boolean
      query:
        type: string
      results:
        items:
          $ref: '#/definitions/services.SearchResult'
        type: array
      search_time_ms:
        type: number
      total_results:
        type: integer
    type: object
  services.SearchResult:
    properties:
      chunk_id:
        type: string
      distance:
        type: number
      document_id:
        type: string
      document_name:
        type: string
      metadata:
        additionalProperties: true
        type: object
      score:
        type: number
      text:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive API for document processing, analysis, and risk assessment
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Risk Analyzer API
  version: "1.0"
paths:
...
  /api/v1/search:
    get:
      description: Perform a simple search using query parameters
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Collection name
        in: query
        name: collection
        required: true
        type: string
      - default: 10
        description: Number of results
        in: query
        name: top_k
        type: integer
      - default: true
        description: Use cache
        in: query
        name: use_cache
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.SearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Simple search
      tags:
      - search
    post:
      consumes:
      - application/json
      description: Perform vector similarity search across documents
      parameters:
      - description: Search request
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/handlers.SearchRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.SearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Search documents
      tags:
      - search
...
  /search:
    get:
      consumes:
      - application/json
      deprecated: true
      description: 'DEPRECATED: Use /api/v1/search instead. Search for similar chunks
        in the vector store'
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: documents
        description: Collection name
        in: query
        name: collection_name
        type: string
      - default: 5
        description: Number of results
        in: query
        name: top_k
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: '[DEPRECATED] Search documents'
      tags:
      - deprecated
...
swagger: "2.0"